Injection and db initialization changes
