+++ pleo-antaeus-app/src/main/kotlin/io/pleo/antaeus/app/Entrypoint.kt
@@ -7,12 +7,9 @@ import dagger.Module
 import dagger.Provides
 import io.pleo.antaeus.core.external.PaymentProvider
 import io.pleo.antaeus.core.services.BillingService
-import io.pleo.antaeus.data.AntaeusDal
 import io.pleo.antaeus.models.Invoice
 import io.pleo.antaeus.rest.AntaeusRest
 import org.jetbrains.exposed.sql.Database
-import org.jetbrains.exposed.sql.transactions.TransactionManager
-import java.sql.Connection
 import javax.inject.Inject
 import javax.inject.Singleton
 import kotlin.random.Random
@@ -23,20 +20,13 @@ fun main() {
 
 class AntaeusApp {
     @Inject
-    lateinit var dal: AntaeusDal
+    internal lateinit var billingService: BillingService
 
     @Inject
-    lateinit var billingService: BillingService
-
-    @Inject
-    lateinit var rest: AntaeusRest
+    internal lateinit var rest: AntaeusRest
 
     init {
         DaggerAppComponent.create().inject(this)
-
-        TransactionManager.manager.defaultIsolationLevel = Connection.TRANSACTION_SERIALIZABLE
-
-        setupInitialData(dal)
 
         billingService.startCronCharger()
         rest.run()
