+++ pleo-antaeus-core/src/main/kotlin/io/pleo/antaeus/core/services/CustomerService.kt
@@ -4,10 +4,12 @@ import io.pleo.antaeus.core.exceptions.CustomerNotFoundException
 import io.pleo.antaeus.data.AntaeusDal
 import io.pleo.antaeus.models.Customer
 import javax.inject.Inject
+import javax.inject.Singleton
 
 /**
  * Implements endpoints related to customers.
  */
+@Singleton
 class CustomerService @Inject constructor(private val dal: AntaeusDal) {
     fun fetchAll(): List<Customer> {
         return dal.fetchCustomers()
